+++ src/data/pokemon/species_info/gen_1_families.h
@@ -8937,7 +8937,7 @@ const struct SpeciesInfo gSpeciesInfoGen1[] =
             gOverworldPalette_Machop,
             gShinyOverworldPalette_Machop
         )
-        .levelUpLearnset = sMachopLevelUpLearnset,
+        .levelUpLearnset = sMachopPartnerLevelUpLearnset,
         .teachableLearnset = sMachopTeachableLearnset,
         .eggMoveLearnset = sMachopEggMoveLearnset,
         .formSpeciesIdTable = sMachopFormSpeciesIdTable,
@@ -9013,7 +9013,7 @@ const struct SpeciesInfo gSpeciesInfoGen1[] =
             gOverworldPalette_Machoke,
             gShinyOverworldPalette_Machoke
         )
-        .levelUpLearnset = sMachokeLevelUpLearnset,
+        .levelUpLearnset = sMachokePartnerLevelUpLearnset,
         .teachableLearnset = sMachokeTeachableLearnset,
         .formSpeciesIdTable = sMachokeFormSpeciesIdTable,
         .evolutions = EVOLUTION({EVO_LEVEL, 36, SPECIES_MACHAMP_PARTNER}),
@@ -9106,13 +9106,11 @@ const struct SpeciesInfo gSpeciesInfoGen1[] =
             gOverworldPalette_Machamp,
             gShinyOverworldPalette_Machamp
         )
-        .levelUpLearnset = sMachampLevelUpLearnset,
+        .levelUpLearnset = sMachampPartnerLevelUpLearnset,
         .teachableLearnset = sMachampTeachableLearnset,
         .formSpeciesIdTable = sMachampFormSpeciesIdTable,
         .formChangeTable = sMachampPartnerFormChangeTable,
     },
-
-    #define MEGA_ABILITY(ability) { ability, ability, ability }
 
     [SPECIES_MACHAMP_MEGA_G] =
     {
@@ -9139,7 +9137,6 @@ const struct SpeciesInfo gSpeciesInfoGen1[] =
         .levelUpLearnset = sMachampPartnerLevelUpLearnset,
         .formChangeTable = sMachampPartnerFormChangeTable,
     },
-
 #endif //P_FAMILY_MACHOP
 
 #if P_FAMILY_BELLSPROUT